Feature/代码优化 (#86)
* test

* 完成首页界面，接口调试
完成记录界面

* 完善首页广告滑动
修改记录的筛选BUG

* 完成提现成功后的memo自动带入功能

* 增加首页的Loading动画

* 修改充值提现页面的箭头错误

* 修改首页逻辑跳转的BUG
修改跳转到ETO详情

* 修改BUG
还有首页轮播广告

* 修改BUG

* 修改冲突

* 修改一个撤单时候数据不准确问题

* 修改首页跳转逻辑

* 修改更新公告的位置

* test

* 隐藏没有交易的币种

* 迁移swift 4.2 版本

* 解决首页卡顿BUG

* 恢复首页解决BUG的注销代码问题

* 完成1.6.0功能开发

+++ cybexMobile/ViewControllers/Transfer/TransferCoordinator.swift
@@ -242,7 +242,7 @@ extension TransferCoordinator: TransferStateManagerProtocol {
                                                                           to_user_id: Int32(getUserId(to_account.id)),
                                                                           asset_id: Int32(getUserId(balance.asset_type)),
                                                                           receive_asset_id: Int32(getUserId(balance.asset_type)),
-                                                                          amount: Int64(Int32(amount)),
+                                                                          amount: Int64(amount),
                                                                           fee_id: Int32(getUserId(fee.asset_id)),
                                                                           fee_amount: Int64(fee_amout),
                                                                           memo: self.state.memo.value,
